chore: extract case registrar
