+++ tests/runner/tests/registry.test.cpp
@@ -16,10 +16,9 @@
 // FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR
 // OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
 // DEALINGS IN THE SOFTWARE.
-//
 /*!
   \file   registry.test.cpp
-  \brief  Unit tests for \ref toy::test::CaseRegistrar and the runner's name-sorted case registry.
+  \brief  Unit tests for the runner's case registry: the duplicate-name scan over a name-sorted list.
 */
 
 #include <doctest/doctest.h>
@@ -32,87 +31,15 @@ void emptyBody(toy::test::Context & context) {
   static_cast<void>(context);
 }
 
-// A registrar stores the pointers it is handed rather than copying the text, so a case's identity in the list is the
-// literal's address. Comparing addresses asserts that contract and keeps the assertions independent of the name
-// comparison the insertion sort itself runs on.
+// A registrar stores the pointer it is handed rather than copying the text, so a case's identity in the list is the
+// literal's address, and the assertions compare addresses.
 constexpr const char * c_alpha = "alpha";
 constexpr const char * c_bravo = "bravo";
-constexpr const char * c_mike  = "mike";
-constexpr const char * c_zulu  = "zulu";
 
 } // namespace
 
-// Registration order does not survive: the list comes out sorted by name. The three register out of order in both
-// directions, so neither a plain prepend nor a plain append reproduces the expected sequence.
-TEST_CASE("test/case_registrar/insertion_keeps_list_sorted") {
-  toy::test::CaseRegistrar * head = nullptr;
-
-  toy::test::CaseRegistrar second{head, c_mike, "m.cpp", 2, &emptyBody};
-  toy::test::CaseRegistrar third{head, c_zulu, "z.cpp", 3, &emptyBody};
-  toy::test::CaseRegistrar first{head, c_alpha, "a.cpp", 1, &emptyBody};
-
-  // Walking with a cursor keeps a broken link a reported failure: dereferencing a chain would crash the case instead,
-  // and the built-in runner has no signal handler to name the case on a console target.
-  const toy::test::CaseRegistrar * node = head;
-
-  REQUIRE(node != nullptr);
-  REQUIRE(node->name() == c_alpha);
-
-  node = node->next();
-  REQUIRE(node != nullptr);
-  REQUIRE(node->name() == c_mike);
-
-  node = node->next();
-  REQUIRE(node != nullptr);
-  REQUIRE(node->name() == c_zulu);
-
-  REQUIRE(node->next() == nullptr);
-}
-
-// Inserting in alphabetical order gives the same result as inserting in reverse.
-TEST_CASE("test/case_registrar/order_is_independent_of_registration_order") {
-  toy::test::CaseRegistrar * ascending = nullptr;
-
-  toy::test::CaseRegistrar a1{ascending, c_alpha, "a.cpp", 1, &emptyBody};
-  toy::test::CaseRegistrar a2{ascending, c_mike, "m.cpp", 2, &emptyBody};
-  toy::test::CaseRegistrar a3{ascending, c_zulu, "z.cpp", 3, &emptyBody};
-
-  toy::test::CaseRegistrar * descending = nullptr;
-
-  toy::test::CaseRegistrar d1{descending, c_zulu, "z.cpp", 3, &emptyBody};
-  toy::test::CaseRegistrar d2{descending, c_mike, "m.cpp", 2, &emptyBody};
-  toy::test::CaseRegistrar d3{descending, c_alpha, "a.cpp", 1, &emptyBody};
-
-  const toy::test::CaseRegistrar * left  = ascending;
-  const toy::test::CaseRegistrar * right = descending;
-
-  while (left != nullptr && right != nullptr) {
-    REQUIRE(left->name() == right->name());
-    left  = left->next();
-    right = right->next();
-  }
-
-  REQUIRE(left == nullptr);
-  REQUIRE(right == nullptr);
-}
-
-// Every field survives registration, so a report can name the source location.
-TEST_CASE("test/case_registrar/preserves_registered_fields") {
-  constexpr const char * c_caseName = "core/utils/trim";
-  constexpr const char * c_fileName = "tests/core/utils.test.cpp";
-
-  toy::test::CaseRegistrar * head = nullptr;
-
-  toy::test::CaseRegistrar node{head, c_caseName, c_fileName, 42, &emptyBody};
-
-  REQUIRE(node.name() == c_caseName);
-  REQUIRE(node.file() == c_fileName);
-  REQUIRE(node.line() == 42);
-  REQUIRE(node.body() == &emptyBody);
-}
-
 // Two cases sharing a name are neighbours in a sorted list, and the search finds them.
-TEST_CASE("test/case_registrar/duplicate_name_is_found") {
+TEST_CASE("test::detail/find_duplicate_name/finds_the_adjacent_repeat") {
   toy::test::CaseRegistrar * head = nullptr;
 
   toy::test::CaseRegistrar first{head, c_alpha, "a.cpp", 1, &emptyBody};
@@ -127,8 +54,8 @@ TEST_CASE("test/case_registrar/duplicate_name_is_found") {
   REQUIRE(duplicate->next()->name() == c_alpha);
 }
 
-// Distinct names report no duplicate.
-TEST_CASE("test/case_registrar/distinct_names_report_no_duplicate") {
+// Distinct names report nothing.
+TEST_CASE("test::detail/find_duplicate_name/distinct_names_report_none") {
   toy::test::CaseRegistrar * head = nullptr;
 
   toy::test::CaseRegistrar first{head, c_alpha, "a.cpp", 1, &emptyBody};
@@ -137,8 +64,8 @@ TEST_CASE("test/case_registrar/distinct_names_report_no_duplicate") {
   REQUIRE(toy::test::detail::findDuplicateName(head) == nullptr);
 }
 
-// An empty registry is not a duplicate, and neither is a registry of one.
-TEST_CASE("test/case_registrar/empty_and_single_registry_report_no_duplicate") {
+// A list too short to hold a neighbouring pair reports nothing instead of walking off its end.
+TEST_CASE("test::detail/find_duplicate_name/empty_and_single_registry_report_none") {
   REQUIRE(toy::test::detail::findDuplicateName(nullptr) == nullptr);
 
   toy::test::CaseRegistrar * head = nullptr;
